requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.6.0
blinker==1.4
cachelib==0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==6.7
cloud-init==18.5
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cssselect==1.1.0
discord.py==1.3.3
distro-info===0.18ubuntu0.18.04.1
howdoi==1.2.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.9
incremental==16.10.1
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.5.0
MarkupSafe==1.0
multidict==4.7.5
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyquery==1.4.1
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.23.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.14.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
youtube-dl==2020.3.24
zope.interface==4.3.2
